Correction Entity Price float-> int

+++ src/Entity/Order.php
@@ -80,7 +80,7 @@ class Order
         return $this;
     }
 
-    public function getAmount(): ?float
+    public function getAmount(): ?int
     {
         return $this->amount;
     }
